cli in a good place
